Partially reverted: "Simplified the comparisons between ItemStacks and unmodifiable ItemStacks."
This has broken item stack comparisons at a few places, because the underlying item stacks may in some cases be both CraftItemStacks: Comparing a CraftItemStack to a SKUnmodifiableItemStack will not work, whereas comparing the underlying CraftItemStacks works.
Unlike noted in the first commit, since the underlying item stacks are assumed to not be unmodifiable item stacks themselves, the order in which these item stacks are compared should not matter.