Fixed: The ShopkeeperByName command argument did not take the 'joinRemainingArgs' argument into account.
The issue was that the 'createIdArgument' method is invoked by the base ObjectById command argument constructor before the child constructor is executed. Any arguments setup during the child constructor can therefore not be used during the construction of the id command argument. The solution is now to pass any arguments for the id command argument to the base constructor, which forwards them to the 'createIdArgument' method.