Removed an old update mechanism for sign shops.
If the sign block is not available at the time it is meant to be updated, this would update the sign later, once it is available again.
This mechanism should no longer be required: The sign is already updated whenever it is spawned, and we dynamically spawn and despawn the sign with chunk loads, and dynamically respawn it if it is detected to be missing.