Fixed: The parsing of command arguments failed in cases in which command arguments depend on the context provided by fallbacks of earlier arguments.
Due to some incorrect assumptions, the parsing of command arguments was prematurely aborted whenever a fallback argument did not consume any of the remaining unparsed arguments. We will now restart the parsing in this case to check if the subsequent command arguments are able to successfully parse with the new context provided by the fallback argument.

In order to still fail with a more meaningful error when the fallback argument successfully parsed but did not consume any of the remaining arguments and the parsing eventually fails nevertheless, we now keep track of the fallback's original root parsing exception before the parsing is restarted. This root exception will override any other parsing error that would otherwise be thrown when the parsing of subsequent aguments eventually fails.

Also:
* Minor formatting changes.
* Added a TODO note about argument completion suggestions not working yet for arguments that depend on context provided by fallbacks of earlier command arguments.