Changes to how the plugin internally represents and interacts with data.
A new DataContainer interface hides whether or not we use Bukkit's configuration API to save and load data, provides a unified view on data that may be stored in either a ConfigurationSection or a Map (eg. when the data is located inside a list), and offers more possibilities for extension, such as providing an unmodifiable view on the data.
As of right now (this might change in the future), DataContainer is a slim wrapper around either a ConfigurationSection or a Map: Instead of copying the data, it reads and writes through to the underlying data source.

Most uses of ConfigurationSection have been replaced with DataContainer, such as:
* Loading and saving configuration settings, including the messages of the language file, and config migrations.
* Loading and saving shopkeeper and shop object data. This includes the use by the internal properties API.
* Serialization and deserialization of ItemData and SoundEffects.

The new ConfigData extends DataContainer to additionally provide access to default config values.
Similarly, the new ShopkeeperData and ShopObjectData wrappers extend DataContainer with common convenience methods that were previously located inside AbstractShopkeeper to interact with the shopkeeper and shop object data.

Other related changes:
* Replaced the Property#loadValue and Property#saveValue with more basic #deserializeValue and #serializeValue methods.
* Some Property loading error messages have slightly changed.
* We no longer generate a new unique id if the shopkeeper data is missing the shopkeeper's unique id. Instead, the loading of the shopkeeper fails now in this case.
* AbstractShopkeeper uses ConversionUtils to parse the shopkeeper's unique id now.
* Moved a few utility methods from ConfigUtils to the new DataUtils, and removed a few no longer used utility methods from ConfigUtils related to reading Maps. There have also been a few other minor changes in ConfigUtils.
* The loading and saving methods of shopkeeper offers have been renamed.
* We now log a warning when we encounter a shopkeeper offer with invalid data. Previously, these offers were silently skipped. Also, the error messages related to the loading of shopkeeper offers have slightly changed.
* Simplified ConversionUtils#toString(Object).