Added DataValueContainer.
This is a DataContainer implementation that dynamically accesses an underlying data source via a DataValue.

Added DataContainer#getDataValueContainer(String) to conveniently create a DataValueContainer that dynamically reads and writes to a specific entry within another data container.
Updated the description of DataContainer#serialize() to take the behavior of this new DataContainer implementation into account.