Minor changes to the shop and shop object type constructors.
In order to reduce the number of constructor variations that need to be supported by base shop and shop object types, concrete shop and shop object types are now always required to specify the list of aliases, even if it is empty.
Also, null is no longer a valid argument for the aliases list.