Added settings to simulate the trading sounds of vanilla villagers and wandering traders.
Added settings:
* `simulate-villager-trading-sounds` (default: `true`)
* `simulate-villager-ambient-sounds` (default: `false`)
* `simulate-wandering-trader-trading-sounds` (default: `true`)
* `simulate-wandering-trader-ambient-sounds` (default: `false`)
* `simulate-villager-trading-sounds-only-for-the-trading-player` (default: `true`)

By default, we now simulate the trading sounds of villagers and wandering traders for the trading player.
These sounds are only played if the trading player is in the vicinity of the shopkeeper. Trading remotely will not play these sounds.
Debug output for these simulated villager sounds can be enabled via the existing debug option `regular-tick-activities`.

Internally, it is now possible to register TradingListeners with the TradingHandler that are informed on various events during the trade handling. This was used to implement the new simulation of villager trading sounds.

Minor internal changes and javadoc fixes for TradingHandler#onTradeAborted. The (nullable) current trade instance needs to be explicitly retrieved from the TradingContext now.

The EntityUtils and TimeUtils classes are final now.