Only set the helmet of zombies and skeletons after we have verified that the equipment is available.
Actually, unlike noted in a previous commit, there are currently no entities that return a null equipment here. However, since the API specified this as nullable, we check for this now just in case this changes in the future.
Removed the 'bug fix' note again from the changelogs, since it is not accurate.