The various shopkeeper and shop object save methods have an additional 'saveAll' parameter now.
This parameter is for example used when creating shopkeeper snapshots, and it controls whether the shopkeeper/shop object should try to also save data that would usually be stored externally.
This is not yet used by any shopkeeper or shop object, but may for example be used by the Citizens shopkeepers in the future to also capture the NPC data when a snapshot is created.