Fixed: Timer was not correctly updating its state when being stopped, breaking an assertion when the timer was started again.
The Timer class more strictly checks its expected states now and logs an error with stack trace whenever one of its operations is called in an unexpected state. We only log a (verbose) error in this case instead of throwing an actual exception, because any issues with the timer potentially yielding wrong values are usually less important compared to issues that could be caused by an exception unexpectedly aborting some external operation. In order to avoid spamming the log with errors when an unexpected timer state is retained and also affects all future timer operations, we only log the error once and then disable all future timer state checks.