Fixed: We were not able to detect item stack deserialization issues.
Bukkit only logs an error in this case instead of throwing an exception.
This also resulted in the loss of shopkeeper data when someone tried to revert from one server version to an older one, because the loading of item stacks that were saved on a newer server version is not supported. We are now able to detect these deserialization issues and then shut down the plugin before any data is lost.

Minor refactor in SKShopkeeperStorage to remove one unnecessary try-catch scope.