Fixed: If a shopkeeper failed to load, or later failed to save its state, the shopkeeper id was able to be saved as a normal attribute into the save file.
We now create a shallow copy of the shopkeeper data that we store inside the SKShopkeeperStorage, i.e. the data that we might later write back to disk when we perform a save, before we insert the separately stored shopkeeper id and continue the shopkeeper loading.

The shopkeeper storage data version was bumped from 2 to 3, so that any previously saved shopkeeper ids are automatically removed again.

Also minor refactors: Moved the shopkeeper loading and the retrieval+preparation of the shopkeeper data container into separate methods.