Improved the feedback messages that are sent when a shopkeeper cannot be created at a specific location.
The previously used generic `shop-create-fail` message was replaced with dedicated messages for the different reasons for why the shopkeeper cannot be created.

Also, it is no longer possible to create mob shopkeepers when they would not be able to spawn due to the difficulty being set to peaceful.

Internal changes:
* Replaced ShopObjectType#isValidSpawnLocation with AbstractShopObjectType#validateSpawnLocation, which additionally sends feedback to the player who is trying to create the shopkeeper.

Message changes:
* Slightly changed the default `must-target-container` message.
* Removed `shop-create-fail`.
* Added `must-target-block`.
* Added `missing-spawn-location`.
* Added `spawn-block-not-empty`.
* Added `invalid-spawn-block-face`.
* Added `mob-cannot-spawn-on-peaceful-difficulty`.
* Added `restricted-area`.
* Added `location-already-in-use`.