Fixed: When setting the zero currency item to AIR, the selling and book shopkeepers were not able to derive the price of a trade from the editor when only the high currency price is specified.
The problem has been that TradingRecipeDraft automatically reordered the input items if the first input item is empty. This automatic reordering has now been removed. Instead, we added #getRecipeItem1 and #getRecipeItem2 which reorder the input items if necessary. These methods are now used everywhere where we previously assumed the input items to be in a valid order.

TradingRecipeDraft#isValid accounts for this change and returns true now if any of the input items is non-empty.

Also: TradingRecipeDraft#hasItem2() has been removed and moved into SKTradingRecipe.