Removed various item related utility functions from the Settings class.
Instead, the respective ItemData settings are either used directly, or via the new Currency and Currencies classes, which are now used to represent currency items.
Internal preparations to support more than two currency items.
The 'low' currency is now referred to as the 'base' currency. This affects for example the "giveCurrency" and "setCurrency" commands.

Message changes:
* Changed currency-items-given.
* Changed currency-items-received.
* Changed currency-item-set-to-main-hand-item.
* Removed high-currency-items-given.
* Removed high-currency-items-received.
* Removed high-currency-disabled.
* Removed high-currency-item-set-to-main-hand-item.
* Added unknown-currency.