The ItemData serializer can now be configured to serialize the display name, lore, and loc name in the plain text format with color codes.
The plain text format is only used if we can verify that deserializing the text would produce the same Json formatted text again.
However, this guarantee is only provided for the currently used server version: If the server makes changes to how it converts from the plain format to the Json format, deserializing the ItemData on a different server version might yield a different result. For this reason, we currently use this setting primarily in test cases in which we either prefer to specify item text data in the simpler plain format with color codes, or in which we compare serialized ItemData with the contents of the default config, in which we prefer to use the plain text format as well.

This change fixes various test cases that broke with the recent upgrade to Bukkit 1.16.x.