Reverted: When interacting with the shop creation item, we no longer ignore interactons with some types of blocks.
This did not properly work before, because we always cancel interactions when holding the shop creation item, so these blocks could not be used anyway.
To reduce the risk of bugs or unforeseen Minecraft logic allowing players to use the shop creation item like normal, we prefer to keep the behavior of always canceling interactions when players hold the shop creation item.

We still ignore interactions with these special types of blocks when we ask players to right click a new shopkeeper location.