Omit the warning about invalid living entity types when we initialize the default config settings
We still log the warning for the initially generated default config
setting when we run on an older MC version. However, unlike before, admins
are now able to resolve the warning by manually adjusting the generated
config to no longer include mob types that are not yet supported by their
server version.