Add shulker block attachment and moving between sign post and wall sign
Shulkers are now oriented according to the block face they are placed
against.

Moving shopkeepers takes the clicked block face into account now. This is
for example used by sign and shulker shopkeepers which now change their
orientation accordingly. However, changing the orientation of a sign or
shulker shopkeeper in place, without first moving it to another location,
is not yet possible.