Disable the gravity of shulkers
If the shulker is oriented to stand on top of a block, we still adjust its
spawn location downwards by up to one block, as we do with any other mob
even if gravity is disabled.
However, we no longer adjust the spawn location if the shulker is oriented
in any other direction.