Fix #809: Error when deleting a Citizens shopkeeper
The issue was that the removal of the Citizens NPC triggered another
deletion of the shopkeeper, which the original deletion procedure does not
account for.

We now mark the shopkeeper as 'invalid' earlier already and skip the
shopkeeper deletion during the NPC removal if the shopkeeper is no longer
valid.