Improve support for hex colors in messages and shopkeeper names
We support both the Bukkit hex syntax ('&x&a&a&b&b&c&c') and the more
compact hex syntax ('&#aabbcc') now.
The Bukkit syntax already worked before for shopkeeper names and some
texts, but is supported for more messages now.

When a player renames a shopkeeper using the compact hex syntax, the
player's input is expanded to the Bukkit syntax due to compatibility
reasons (Spigot only supports the expanded hex syntax in certain contexts).
Validations like the shopkeeper name length limit are applied to the
expanded syntax.

Note: In order for hex colored shopkeeper names to properly work in the
trading UI, you need to use the latest Spigot 1.19 version or newer.

Internal: Text.toPlainFormatText() has been renamed, and changed to
represent color and formatting codes using the '&' character.