Add command "/shopkeeper replaceAllWithVanillaVillagers"
This utility command deletes all shopkeepers and replaces them with
corresponding vanilla villagers without AI that are configured very similar
to villager shopkeepers.

This might for example be useful when migrating a world to vanilla
Minecraft, e.g. when a server closes a world but wants to provide it as
download to its players with all the shopkeepers included.

This command requires all of the following permissions:
* `shopkeeper.debug`
* `shopkeeper.remove-all.player`
* `shopkeeper.remove-all.admin`