Revert: Remove setting `add-tag-to-shop-creation-items-in-shops` again
In most cases, except for maybe admin shops and certain hire cost setups,
it makes no sense to automatically migrate the shop creation item in the
shop data without also having a solution to automatically migrate the items
in the world data e.g. in chests, player inventories, etc. Applying the
migration only to the shop data breaks existing trades and hire options and
prevents players from trading in their old shop creation items that they
might still possess. So the existence of this setting does not provide much
benefit and might actually cause issues when used without careful
consideration.

In the few cases in which this setting might make sense to use, the effort
to one-time manually update trades or hire cost items is expected to be
minimal, and can for example be achieved via a simple text replacement in
the save file.