Avoid adding empty item stacks to merchant recipes
In the past, this was done to mimic how Spigot populates the recipe
ingredients and to simplify merchant recipe comparisons. With 1.20.5, this
is no longer being done by Spigot, and results in an error on Paper due to
more stricter checks that have been added there.

Merchant recipe comparisons were adjusted to filter out empty item stacks
before comparing ingredients.