Avoid extending Bukkit's ItemStack
Letting our custom SKUnmodifiableItemStack extend Bukkit's normal ItemStack
and overriding all modification methods to throw an exception instead had
the benefit that we could pass the SKUnmodifiableItemStack to any methods,
internal or external (e.g. in the Bukkit API itself), without having to
copy the item stack first, while still ensuring that the underlying item
stack is never modified, even if it is accidentally passed to a method that
was originally not expected to modify the passed item stack.

However, this approach had a few disadvantages:
* The Bukkit API does not officially support this approach and this case is
  unlikely to be taken into consideration when evaluating backwards
  compatibility concerns.
* Whenever the Bukkit API adds new methods to the ItemStack class, we need
  to update our implementation, either to delegate to the underlying
  ItemStack, or to throw an exception.
* The implementation of SKUnmodifiableItemStack#equals(Object) was not
  symmetric because the server's CraftItemStack does not consider custom
  ItemStack implementations as equal.

Additionally, the Paper server for MC 1.21 recently made some changes to
ItemStacks that also do not take custom ItemStack implementations into
account and thereby broke the Shopkeepers plugin.

In order to resolve those issues, we no longer extend Bukkit's ItemStack.
Instead, whenever we need to pass an ItemStack to a method, internal or
external, that we know/expect to not modify the passed item stack, we now
pass the underlying ItemStack instead. This is slightly more risky, because
we are no longer notified, in the form of an exception, if the operation
ends up trying to modify the passed item stack. E.g. on custom server
implementations, there might sometimes be discrepancies that we would
usually want to detect and then report.

This optimization of accessing the underlying item stack is not exposed by
the Shopkeepers API and only used internally by the Shopkeepers plugin
itself. Add-on plugins currently always need to copy the item stack when
they need to pass an ItemStack, even if they know that the receiving method
does not modify the item stack.

In some cases, when converting from an unmodifiable item stack to a normal
item stack and then back to an unmodifiable item stack, we can end up
creating more unmodifiable item stack wrapper instances now, because the
optimization of just returning the item stack if it already is unmodifiable
no longer works. But this is a minor issue, because the wrappers are
lightweight objects.

In some cases, we might end up unnecessarily copying item stack data now,
because we can no longer identify if a given Bukkit ItemStack is already
unmodifiable.

Also:
* Mark some more methods that accept a normal ItemStack as @ReadOnly.
* API: Add PlayerShopkeeper#setForHire(UnmodifiableItemStack). This might
  break some add-on plugins, because when passing null, the call is now
  ambiguous and requires a cast to either ItemStack or
  UnmodifiableItemStack.