Add the option to add default potion effects to living entity shops
The default effects are preserved and periodically re-added. Any
non-default effects are periodically removed, or even prevented from being
added in the first place.

This feature is not yet used. It might for example be used in the future if
a certain mob type requires a certain potion effect to properly function as
a shopkeeper.