Revert: Immediately substitute placeholder items in the equipment editor
This has the benefit that the equipment does not break should the
placeholder item no longer be recognized in a future server or plugin
version.

For the carried block of enderman shopkeepers, if the block has no
corresponding item type, we still substitute the placeholder item
dynamically whenever the carried block is applied.
For blocks that have a corresponding item type, we now preserve the block
state data present on the placeholder item on the substituted item, so that
we can still apply this data to the carried block of endermans.

Add clarifications.